Add test for tag-release.sh using rspec/aruba
Add a test around tag-release.sh, using the helpers and library for
aruba on rspec.

This test prepares a fake origin repository with tags, for which it uses
the helper bash_c to execute a batch of shell commands using aruba, and
mocks the state the script will find in concourse.

The script had to be changed to adapt it to the test and to be able to
run test it locally and in travis, like add the option of defining an
alternate remote url, or use bash rather than sh.