Allow users to switch backend
Because of cf-acceptance-tests issue 104
(https://www.pivotaltracker.com/n/projects/1382120/stories/117685687)
the multiple app ports test fails if users cannot select backend.
We enable this for the test to pass, and it has no bad consequence as
if the user tries to switch backend, it would fail as we only have
diego nodes.