Set route_registrar routes to empty array
route_registrar package is required for statd_injector to function
correctly because it provides the 'common' package dependency required
by statsd_injector - however we do not actually use this component
anymore as we moved the routes over to ELBs. Setting the `routes`
array to empty allows the route_registrar job to start successfully
without it creating any routes.