Disable the 'buildpack from a git url' test
The acceptance test 'uses a buildpack from a git url'[1] from the
`internet_dependent` suite fails if we disable the custom buildpacks for
end users, which was expected.

We disable the test by name, and we will add one to check that the
feature indeed fails in the custom tests.

[1] https://github.com/cloudfoundry/cf-acceptance-tests/blob/f39ba990159dee932e99a6ca829df79a42746822/internet_dependent/git_buildpack_test.go#L20