Extract update-ca-certificates to script
We repeat this pattern several times in the pipeline and I'd like to re-use
it in order to remove the `--skip-ssl-validation` arguments from `cf login`.
So I've extracted it to a script which can be called in the pipeline.

It expects `bosh-CA/bosh-CA.tar.gz` to be available. The `set -e` will cause
tar to abort the script with a non-zero exit code if the file isn't present.