Remove --skip-ssl-validation from cf login
By importing the BOSH CA, which currently signs our public facing
certificates, we can re-enable SSL validation and be sure that we are
talking to the real CF API.

This won't be necessary for staging and production when we have real
certificates issued by a third-party CA. However we'll still need it for
development and it probably causes no harm to have both CAs installed.

I've combined the `cf api` and `cf login` commands because I can't see a
reason for these to be separate.