Add port-scan job
Run a port scan concurrently with other acceptance tests. Do a staging
and run-time scans in parallel.

The default staging app timeout is 15 mins, which is enough for this
scan, as with 39 VMs it finishes in about 5 mins. Should this scan get
considerably longer, the app would fail staging and the task would fail
as well - indicating that there are some problems on the platform.

If we desire to increase the scan scope in the future, the staging timeout
can be adjusted to be longer.

The runtime scan runs in background to the python app, which starts
without any delays. The job then polls scan status and retrieves results
when scan finishes.

The job then checks for open ports. You can specify if you want to see
the scan results always or only when open ports are found.