Ensure etcd cluster is updated serially
Without this, when upgrading the cluster (eg for a stemcell version
change), the first node will be upgraded, then the other 2 will be
upgraded in parallel, resulting in the cluster losing quorum. This then
requires manual intervention to recover from.

Making etcd_z2 serial as well ensures that one node gets done at a time.