Fix bootstrap pipeline.
0545bbe updated this pipeline so that it would work with newer versions
of concourse and pass the linter. This fix wasn't complete because the
added outputs created empty directories that were no longer being
populated, meaning that later tasks couldn't find the necessary files.

This updates the pipeline to be closer to our styleguide and have output
names that are distinct from the task names. It is then explicit in
what's added to the output directorties.