cf-terraform does not need bosh-tfstate
The cf and bosh terraform jobs have run in parallel since #213 was
merged. This suggests that the cf-terraform job doesn't depend on the
outputs from bosh-terraform. Reading the code confirms this - there are
no variables in the cf-terraform that match the names of outputs in
bosh-terraform. Additionally the cf-terraform destroy job doesn't rely
on the bosh-ftstate file.

This therefore removes the bosh tfstate from the cf-terraform job.