Upload and use paas-haproxy in front of go-routers
Upload the custom paas-haproxy release as a development release as we
do with other releases.

paas-haproxy will allow us to define additional raw haproxy configuration
in the manifest to add additional features like setting headers.

We add it to the manifest to configure colocated it in front of the
go-routers, for which we must:
 * Add it as the list of releases for the deployment
 * Disable SSL listeing in the go-router, as haproxy will do the SSL
   termination. This way go-router won't listen in port 443.
 * Configure HAProxy:
   * Unique backend will be the go-router in localhost (colocated)
     without SSL
   * Will listen HTTPS with the required certificate
   * Will NOT listen HTTP. go-router will.