Split system and apps external certs.
When we came to buy real certs for production and staging, it wasn't
feasible to get a single sert that covered multiple wildcard domains.
We've therefore had to get separate certs for the system and apps
domains.

This updates the dev/ci cert generation to also generate separate certs,
and updates the pipelines to use these separate certs.