val_from_yaml: Test command line interface
Refactor the existing tests for `val_from_yaml` to use the command line
interface instead of the underlying Ruby object.

This matches how we use the script in our Concourse pipelines, it exercises
more of the code, and allows us to make assertions about the standard output
and exit codes.

I'm using heredoc strings for the results because they must contain a
trailing newline. I have used a single fixture file because the contents
don't change between tests and it negates us writing a temp file each time.

The exit code for the failure cases (nil/empty value) is incorrect. I'll fix
that in a subsequent commit.