Add ./scripts/create-tenant.sh script
./scripts/create-tenant.sh will create a user and organisation in the CF
API where you are currently logged in and send an email to the user
if the password changes.

The script leverages the cf client to create the organisation, space
and user, and use AWS cli with SES to send emails.

The user password can be reset by deleting and creating the user again
although this has the drawback that the user will be removed from
all the organisations that they were.

The template of the subject and the body, the origin IP and default
space to create are defined in the top of the script.