Remove logs from alert email
The commit message and smoke test logs are useful information to send
with the alert email but they can contain characters incompatible with
json or aws command line.
We now remove them from the email so we can be notified.

The logs generation is not removed so we can send them later with a
more robust approach.