Add Rubocop lint checks
Rubocop is a static code analysis tool for Ruby. It comes with a bunch of
checks, called "cops". Some of them, such as the "style" cop, are very
opinionated and have limited value.

However I've found the "lint" cop very useful for catching unintended
behaviour in previous projects. It is described as:

> Lint cops check for possible errors and very bad practices in your code.
> RuboCop implements in a portable way all built-in MRI lint checks (ruby
> -wc) and adds a lot of extra lint checks of its own.

We now have enough Ruby code, by way of scripts or tests, that this is worth
adding. By adding it to the `test` make target it will be run by Travis CI
(and locally) alongside our other tests. I'll fix all existing violations in
following commits.