Hardcode PASSWORD_STORE_DIR for Terraform scripts
These weren't working quite as desired due to a combination of GNU Make
oddities and refactoring. I noticed while reviewing #236, but I've only just
got round to looking at it.

manually_upload_certs
---------------------

Expansion of `~/` in `CERT_PASSWORD_STORE_DIR` worked when specifying it as
an environment variable:

    ➜  paas-cf git:(master) ✗ CERT_PASSWORD_STORE_DIR=~/.paas-pass-high make staging manually_upload_certs
    …

But not when passed as an "argument" (I'm not sure what the right is -
variable?) to `make`, as described in our documentation:

    ➜  paas-cf git:(master) ✗ make staging manually_upload_certs CERT_PASSWORD_STORE_DIR=~/.paas-pass-high
    #FIXME: Remove this when tag verification has been fixed
    # check password store and if varables are accesible
    Error: certs/staging/staging/system_domain.crt is not in the password store.
    make: *** [manually_upload_certs] Error 1

It also seems that the default had no effect after the refactoring to move
it to a separate shell script in 0262322:

    ➜  paas-cf git:(master) ✗ make staging manually_upload_certs
    #FIXME: Remove this when tag verification has been fixed
    # check password store and if varables are accesible
    terraform/scripts/manually-upload-certs.sh: line 5: CERT_PASSWORD_STORE_DIR: unbound variable
    make: *** [manually_upload_certs] Error 1

pingdom
-------

The `eval export PASSWORD_STORE_DIR` in the `Makefile` had no effect because
`PASSWORD_STORE_DIR` was being explicitly exported within the shell script.

Even if the explicit export is removed, it still doesn't work as I would
have expected, because it also suffers from the same problem with `~/` not
being expanded.

Solution
--------

Hardcode both of these paths within the scripts that use them to
standardised paths. This means that you can no longer have your password
store in a different path, but I think that's OK.

I think there might have been some rationale in having
`CERT_PASSWORD_STORE_DIR` be a different variable name from
`PASSWORD_STORE_DIR` which `pass` uses. But since we're no longer passing
anything in, I think that's also OK.

I have also omitted the check in the `Makefile` and instead put an `echo`
statement in both scripts that states where it's using the password store
from. This, combined with the `is not in the password store` error message,
should help you figure out what is wrong if you don't have the password
store cloned in the right place.