Acceptance test for Db instance creation and connection.
Tests the full lifecycle of a broker DB instance.

DB creation and destruction happens asynchronously, so we've had to poll
the API here. In order to prevent this polling adding unhelpful amounts
of output to in verbose mode, This includes a quietCf function that
doesn't send the command calls, and output to GinkgoWriter.