Add terraform outputs and use in manifests
The security group that allows connections to the db instances is
applied to the rds broker and the cells. The security group that db
instances are placed in are added to the broker config so it will create
dbs in that security group. The other configuration for the service
broker is updated so it can create rds dbs in the correct subnet group.