Use ruby container to extract TF veriables in destroy
This task has no need to use the bosh-init container (it's running the
same extract scripts that are run elsewhere in the pipelines in the
plain ruby container). Switching this to the ruby container makes it
consistent with other jobs in the pipelines, and also means it uses a
smaller container.