Force rds-broker catalog refresh on deployment.
If the rds-broker is deployed with changes to the service/plans, these
weren't being picked up by the cloud-controller. The docs[1] say that
the catalog is queried when the broker is intially registered, and also
when it's updated, but make no mention of any polling for changes to the
catalog.

This therefore adds a `cf update-service-broker` call if the broker is
already registered to force a refresh of the catalog.

[1]https://docs.cloudfoundry.org/services/api.html#catalog-mgmt