Turn off V3 API tests, which don't currently work
The V3 suite of tests currently fail when presented with HTTPS endpoints and
there doesn't seem to be any way around this.

Reported upstream: cloudfoundry/cf-acceptance-tests#110

Since we're not yet using the V3 API it seems safe to turn off the tests for
it.

Additionally...

The acceptance-tests suite for v233 includes a test for the v3 api
which fetches a go buildpack from github. Within the last 48 hours,
the go buildpack has been changed in github in a manner which
appears to make it incompatible with the version of the tests
we're running. Later versions of the acceptance tests have been
swtiched to use a ruby buildpack instead of the go buildpack

We need to disable the v3 acceptance tests (which are disabled
already in the PR for upgrading to CF 235) to unblock the pipelines.