Split endpoint access cidrs by role type
We weed to whitelist access to CF API and Web Access to apps to the
tenants out of the GDS office in the early access to the platform.
This adds a new kind of role to our policies.

In order to simplify the access rules and definition of CIDRS,
we split the endpoint cidrs list definitions for the security groups
in three types of roles:

 * admin_cidrs: Access to everything, from operation endpoints to CF API
   and web access to apps.
 * tenant_cidrs: Access to CF API and web access to apps.
 * web_access_cidrs: Web access to apps.

These variables can be overwritten in the terraform/*.tfvars, for
instance in terraform/prod.tfvars