Use compact whenever we splot admin_cidrs
The function compact[1] will clean up any empty string from a list of
string. This is really useful in the that admin_cidrs is a empty string,
as in that case we want split by comma return an empty list.

We add compact to all the places where we split the admin_cidrs to keep
consistency with other parts of the code.

[1] https://www.terraform.io/docs/configuration/interpolation.html#compact_list_