Disable verbose output for platform tests
It makes the test output more difficult to read, in both success and failure
scenarios, because you have to scroll through much more content. Prior to
af323b6 it was also causing the admin password to be output on every test
run.

This was originally enabled so that we can see the reports from performance
tests. After the preceding commit it is no longer needed for that purpose.

Disabling it does come at the cost of not being able to see in realtime how
the test setup is progressing. I think that's OK though, since our own
suites don't take that long and it matches the behaviour of the jobs for the
upstream tests - cf-acceptance-tests / cf-smoke-tests.