Add trade tariff quota with 60GB
Add the specific quota for the trade tariff quota. Based on
https://graphite.publishing.service.gov.uk/dashboard/temporary-3#temporary-3
one trade tariff environment uses less than 15GB memory. We can cap
at 20GB for safety, therefore 60GB for 3 environments.

In the future we will create more generic quotas, as soon as we have
defined our quota model.

In this commit we remove the `meta` section in the manifest, as this
configuration is not reused in the manifest.