Support for multiple quotas in update-quota tasks
We want to update all the quotas defined in the manifest, not only
default.

To do so, we update the script extract_quota_settings.rb so it will
read the manifest and generate one file per quota with all the required
variables as shell script.

Then, the update-quotas task will read each file and update the quotas
accordingly. We use a subshell[1] to ensure that we only use the
environment variables from each quota.

We also add tests for the ruby script extract_quota_settings.rb.

[1] http://www.tldp.org/LDP/abs/html/subshells.html