Refactor custom tests layout.
Make them use a common src directory, and move all the example apps into
a single directory.

This means that we can setup the GOPATH correctly (directories in the
GOPATH are expected to contain the `src` directory), which it wasn't
before. This will fix the warnings we're currently getting when running
godep restore.