Force delete in destroy-cf pipeline
Boyscout: We already do force delete in CF autodelete, we should
also in regular CF destroy pipeline. `--force` option solves problems
like VMs unable to complete drain script (because other half of
deployment is gone) or when you try to remove remains of the deployment
that is in weird state because you were testing something.