Reduce the length of the temp user for smoketests
In the continous smoke tests we were creating a user called
`continuous-smoketest-user`. With the random suffix that makes
a user name of 36 chars.

But it turns out that the deployment of applications fails when
using a user with 36, 35 or 34 characters in the name only in prod,
but works on staging and ci.

The error message presented during the smoke tests is seen after the tests
attempt to `cf push` the sample ruby application. Instead of the normal
'1 of 1 instances running' text being displayed, the following error is
shown:

  Could not fetch instance count: Server error, status code: 503,
     error code: 200002, message: Stats unavailable: Stats server
     temporarily unavailable.

Longer user names seem to work. We still need to investigate the root case
in a different story.

In order to allow the continuous smoke test work on prod, we change
the username to a shorter one.