Fix bosh_login.sh when password starts with '-'
If the random generated password for the bosh admin starts with '-', the
`bosh_login.sh` won't work because it recognises the password with an
option.

To avoid that we add '--' to tell bosh that the following parameters
are not options.

Additionally, if the string starts with '-', ruby yaml will quote the
string in the cf-secrets.yml with `"`. We need to remote these quotes
before passing the password as argument.