Clarify broker and service naming.
When creating a service broker in cloudfoundry it needs to be given a
name.  This is only used to refer to that broker itself, and isn't
exposed to tennants, instead they use the service name. This had got
confusing because the broker had been given the same name as the service
it provided.

This therefore renames the broker to rds-broker to clarify this
situation.

Also update the constant in the acceptance test to clarify what it's
referring to.