Ginkgo suite for "run_from_vm"
Add a new test suite that will need to be run from a BOSH VM rather than a
Concourse or Diego container.

These were created with:

    ginkgo bootstrap
    ginkgo generate mtu
    godep save

Right now the easiest way to run this is to compile a binary of the tests:

    GOOS=linux ginkgo build

Then transfer that binary to a VM (somehow) and run it with:

    API_ENDPOINT=https://api.<system_domain> ./run_from_vm.test

Later on we can figure out how to automate this in the pipeline. Passing
configuration as environment variables could change then, but it's the
simplest thing for now.