tests/run_from_vm: Requests of different sizes
We experienced a bug in CF on AWS whereby TPS was unable to query Doppler
with an `Authorization` header that contained a username of a specific
length.

This test reproduces the same behaviour by making HTTP or HTTPS requests
with increasingly large packet sizes near to the standard network MTU of
1500 bytes. It should prevent the same from happening again in the future
without us needing to debug the internal CF calls.

It needs to be run from a VM that is configured with an MTU of 9000~ (aka.
Jumbo Frames) and against an endpoint that is accessed via a NAT Gateway and
ELB. It cannot be run from a garden-linux container (eg. Concourse or Diego)
because the virtual interface presented to the container has an MTU of 1500
regardless of whether the host is configured to use 9000.

When run against CF's API endpoint the test currently fails at a request
size of 1397 bytes. It doesn't continue on failure, but if it did, it would
also fail with 1398~1400 and then succeed at 1401 or more.

The request sizes described above and specified in the test aren't the same
as packet size seen by tcpdump. This is because of the overhead added by
IP+TCP and TLS (if in use). We can't compute the exact size without fiddling
with raw sockets, so we try a range of sizes instead.

The range of sizes needs to be tested from small to large and halted on the
very first failure because we've seen the large requests can fix the problem
for a temporary period of time.

Hector wrote a TCP echo server which could also test this behaviour without
fiddling with HTTP. However it would have been harder to integrate as a test
in our pipeline because it would have required running the server code
somewhere and configuring an ELB, whereas this test uses our existing
endpoints.

- https://gist.github.com/keymon/907ef1f942ee532a894c3fabe0aa03de