tests/run_from_vm: Check MTU of host
Test that the suite is being run from a host with a network interface
configured with an MTU of 9000~, because that is needed to exercise the bug
found by the other test in this suite.

I didn't want to hardcode the interface name/index because it wouldn't be
portable between operating system types (OSX and Linux), versions (udev
random naming), or containers. This required quite a bit more code though,
because it needs to connect to the endpoint in order to find out which local
interface the connection is routed through. It also requires two steps
because we get an IP address and then have to find the interface it belongs
to.

I used `http.Client` to make the connection and hijack the `net.Dial`
functions because `net/http` has existing logic for converting the
`http`/`https` schemes to the right port number whilst respecting any `:n`
port at the end of the hostname.

I'm allowing a variance of 10 when checking the MTU value because it's quite
common to use a value of `9000` but AWS uses `9001`.