val_from_yaml: STDERR msg when unable to find key
Write an error message to STDERR when `val_from_yaml` is unable to find a
key in the YAML document. This will make it easier to diagnose typos when
running from within the pipeline. Previously the non-zero exit code would
make the task fail without any output.

For example, when called as we do in the pipeline:

    ➜  paas-cf git:(master) ✗ TEST=$(./concourse/scripts/val_from_yaml.rb foo manifests/concourse-manifest/concourse-base.yml)
    Unable to find key: foo