Refactor quota setting task
This was implemented in 2 separate places - a ruby script to extract
details from the manifest, and an inline shell script in the pipeline to
apply these details. There was also a bug whereby it couldn't add a
quota that didn't already exist (cf update-quota errors if the quota
doesn't exist).

This refactors it into a single ruby script that reads the manifest and
applies the quotas. This now follows the same pattern as the task to set
the security groups.