destroy-microbosh: Abort if any deployments exist
It's possible to run the `destroy-microbosh` pipeline before the
`destroy-cloudfoundry` pipeline. When this happens, all of the Cloud Foundry
VMs and their related EBS volumes are left "orphaned". You have to cleanup
manually using the AWS console, which is tiresome and prone to error.

Prevent this from happening by aborting the `destroy-microbosh` pipeline
before it calls `bosh-init delete` if there are any existing deployments.
The script lists the deployments that you need to delete.

To test this I've used a library called `mimic` to mock the minimal API
responses from BOSH and overridden `BOSH_CONFIG` to point `bosh_cli` at that
fake BOSH.

I looked at using `webmock` as we've used elsewhere, but it only works for
HTTP calls within the same Ruby interpreter and I want to test the script in
a different process using Aruba so that we can assert on the exit code and
STDERR. I looked at the library `webmock-server`, which allows you to
configure a daemon using normal webmock stubs, but it looked more
complicated and could have affected other tests in this suite.