Add a ruby version file
This matches the ruby version specified in the .travis.yml file

The [ruby
version](https://rvm.io/workflow/projects#project-file-ruby-version)
file is supported by multiple ruby version managers and allows changing
the version of ruby to the correct one automatically.

I added this because on working on #118838943, we added a gem which
would not install with the older version of ruby I was running.