Move global properties into cf-properties
Having some of these defined in cf-properties and some in cf-jobs is
confusing as it's not clear where to look for them, or where new
properties should be added. Putting all of these in the same place makes
it clearer.

This also moves the meta.diego_templates details into cf-jobs where
it's referenced from.