Add new quota for test applications
The health check application is to be replaced with one
that serves a static endpoint and a database endpoint.
This means the test application will need to be deployed
with a database service instance, and the only service plans
available are paid ones. Adding this quota allows the test
application to create a service instance, without affecting
the services and plans available to our tenants.