Force IPv4 loopback for bosh_pre_destroy test
We've seen an intermittent problem with the `bosh_pre_destroy.rb` tests
running in Travis CI whereby BOSH is unable to connect to the fake BOSH
server that's provided by Mimic.

The error is truncated in the test output because Aruba's command timeout kicks
in before BOSH has retried 4 times and printed the traceback. But you can see
it by setting `aruba_config.exit_timeout=60` in `support/configure_aruba.rb`:

     1.2) Failure/Error: raise Exception.new("Last command output: " + last_command_started.output)

          Exception:
            Last command output: [WARNING] cannot access director, trying 4 more times...
            [WARNING] cannot access director, trying 3 more times...
            [WARNING] cannot access director, trying 2 more times...
            [WARNING] cannot access director, trying 1 more times...
            /home/travis/build/alphagov/paas-cf/vendor/bundle/ruby/2.2.0/gems/bosh_cli-1.3202.0/lib/cli/client/director.rb:856:in `rescue in perform_http_request': cannot access director (Connection refused - connect(2) for "::1" port 25555 (::1:25555)) (Bosh::Cli::DirectorInaccessible)
            …

Each time it fails it can be seen trying to connect to the IPv6 loopback
address. When logging is added it can be seen that all of the successful
runs use IPv4 loopback instead.

It seems that IPv6 is disabled in Travis' container based infrastructure
(which is where our tests are running):
https://docs.travis-ci.com/user/ci-environment/#Networking

> The virtual machines in the Legacy environment running the tests have IPv6
> enabled. They do not have any external IPv4 address but are fully able to
> communicate with any external IPv4 service. The container-based, OSX, and
> Trusty builds do not currently have IPv6 connectivity.

So Mimic won't be binding dual-stack to `::1`. travis-ci/travis-ci#4978
describes that there are incorrect `hosts(5)` entries for IPv6. Though it's
not clear why BOSH sometimes resolves `localhost` to IPv4 and sometimes to
IPv6.