Start fake BOSH once in bosh_pre_destroy tests
We start it once at the beginning of the suite, add the `/info` route before
each test, remove all routes each test, and then shutdown the server at the
end of the suite.

This is instead of starting a new server at the beginning of each test and
stopping it at the end of the test, which was potentially slower and had a
race condition whereby the previous server hasn't released the port before
the new server starts.