Updated healthcheck app deployment in pipeline
The health check application has been replaced with a database
application serving a static endpoint and a database endpoint.

The `cf create-service` command is asynchronous, so we have to wait
until it is created before binding the service and starting the
application.

There is currently no tracking of state when deploying using CF CLI
in the pipeline. `cf create-service` and `cf bind-service` both exit
0 and do nothing if the service already exists and is bound.

The task was moved out of the aggregate block to make sure it runs
after the creation of the RDS broker, on which this task is dependent.