Add runtime manifest to set MTU
We have discovered an MTU related problem with AWS's nat-gateway service
which is causing us all sorts of problems at the moment, including:

* 'Error tailing logs: Unexpected EOF'. Seems to be that doppler cannot
properly validate the auth token with the  UAA (presumably because
it's using the UAA endpoint URL and that's being routed out via the
aws-nat-gateway)
* 'Stats unavailable: Stats server temporarily unavailable.' Seems to be
a problem with tps-listener failing to connect to the doppler endpoint

The MTU is different for different machine sizes on AWS, but a lot of
the larger ones (which we're using) now default to using Jumbo frames,
with MTU 9001. AWS doesn't give us enough control over the
dhcp-option-sets to set the MTU for ourselves.

The problem doesn't happen if you spin up your own nat-masquerading
linux box. AWS has confirmed they see the problem when we sent them some
code to try out MTU sizes from 1513 to 1517 bytes. They say the've worked
out a fix and will deploy it by 30th June.

In the meantime we can set the MTU to 1500 to workaround the issue.
However this is set by DHCP and bosh-agent seems to overwrite the
dhclient.conf when it starts up. But we can use a bosh addon to write a
DHCP hook to set the MTU whenever the lease is renewed.

Once the AWS fix is deployed and we have tested it successfully, these
MTU runtime changes can be reverted.