Use per-account email addresses for acc creation
We do not have the same email addresses SES-validated in all
environments, preferring addresses that are +tagged with the environment
name so they can be easily filtered.

This also removes the default source email address for the mail helper
so specifying one is mandatory. It also removes DISABLE_USER_CREATION
and uses the presence of an email address in the env var to
enable/disable the account creation, matching the behaviour of
ALERT_EMAIL_ADDRESS.