Create and use a Parameter Group for RDS broker
In order to enable the `ssl` and `rds.force_ssl` options for the
instances created by the postgres RDS broker, we create a new parameter
group with the options set in terraform and define it in the broker plans.

We must set them with `apply_method = "pending-reboot"` because they
are static parameters which require reboot. Otherwise terraform
might fail during update with:

    * aws_db_parameter_group.rds_broker_postgres95: Error modifying DB Parameter Group: InvalidParameterCombination: cannot use immediate apply method for static parameter static