Fix find_diverged_forks crash when branch!=master
In cases where the default branch (of either our fork or the origin) is
not `master` we will crash as the comparison is invalid, so use the
current default branch to allow tracking of branch renames and
reconfigurations.

Also includes some slightly better output formatting, and a
requirements.txt for pygithub.