Specify a branch for os-conf-release resource
The `tag_filter` option for the git-resource will by default only look for
tags that resolve to commits on the `master` branch, however the tag that we
were giving it resolves to a commit on the `gds_master` branch of
alphagov/os-conf-release.

This caused Concourse to hang in a rather surprising way. The `cf-deploy`
job wasn't triggered automatically, so at first glance it looked like the
pipeline had run successfully. If you triggered the `cf-deploy` job manually
then it got stuck in a "pending" state with a message:

    waiting for a suitable set of input versions

Tim found that the reporting for this has improved in later versions of
Concourse:

- https://www.pivotaltracker.com/n/projects/1059262/stories/118715899

It's not clear how we missed this in testing. We can look at that after
we've resolved the immediate problem with our CI pipeline.