Fix typo in admin sync logging
When `sync-admin-users` was first run in the staging and prod environments
it deleted a lot of old test users but the logs didn't indicate what their
usernames were:

    I, [2016-05-31T09:26:48.723823 #15203]  INFO -- : Deleting admin user  which is not in the list.
    I, [2016-05-31T09:26:48.840621 #15203]  INFO -- : Deleting admin user  which is not in the list.
    I, [2016-05-31T09:26:48.986202 #15203]  INFO -- : Deleting admin user  which is not in the list.
    I, [2016-05-31T09:26:49.097718 #15203]  INFO -- : Deleting admin user  which is not in the list.
    I, [2016-05-31T09:26:49.219233 #15203]  INFO -- : Deleting admin user  which is not in the list.
    …

This is because it was using a hash key with the incorrect capitalisation;
`userName` instead of `username`. This resulted in a `nil` object which is
cast to an empty string.

This makes the tests pass after the preceding commit to use `Hash#fetch`
made them fail and should prevent this from happening again in the future.