Use Hash#fetch in all admin sync scripts
Make all hash lookups where we're getting (not setting) a value use `#fetch`
instead of `#[]` so that an exception is raised if we mis-type the key name
or the object isn't populated as expected. Previously a `nil` object would
be returned and the scripts could fail in a silent or unclear way.

This change highlights one such case, whereby we're using a key name with
the wrong capitalisation, which I'll fix in a following commit. We don't
explicitly test the log function where it occurs, but it now makes the tests
fail because we exercise the code path and get an exception:

     Failure/Error: get_logger.info("Deleting admin user #{user_info.fetch("userName")} which is not in the list.")

     KeyError:
       key not found: "userName"

I haven't changed the `self.ua.all_pages` calls which return the first item
because it seems that it's expected for them to return `nil` if there are no
results.