Add support for ssh keys to SecretGenerator
This returns an RSA private key, and the corresponding public key
fingerprint.

To generate the fingerprint, I've lifted a method from the sshkey gem
(version 1.8.0). This seemed better than adding a dependency on the gem
(and requiring bundler or similar at runtime etc) purely for this one
method.

This is something that's in the spirit of the Go proverbs[1] - "A little
copying is better than a little dependency."

[1]http://go-proverbs.github.io/