Feature flag destroy pipelines off in non-dev envs
Add a new `ENABLE_DESTROY` feature flag to control whether the
`destroy-cloudfoundry` and `destroy-microbosh` pipelines are setup on the
Deployer Concourse. This is enabled on development environments and disabled
by default on all other environments such as CI, staging, and production.
Instructions for enabling the destroy pipelines in these environments, if
necessary, is added to the non-dev deployments documentation.

This will prevent the destroy pipelines from being run by accident in
persistent environments. We've seen this happen once and it's easy to
do when you have multiple Concourse tabs open to diagnose a pipeline failure
in another environment.

I've not included a warning like `ENABLE_AUTODELETE` does because it will be
quite noisy and the effect isn't as surprising as autodelete; which contains
a pipeline that you either expect to or not to run automatically. I've used
a glob to capture both destroy pipelines in the case statement so that we
don't have to match their exact names.

These pipelines aren't reinstated by `self-update-pipeline` because that
passes in a single pipeline name. I've had to pass the `ENABLE_DESTROY`
environment variable to the `self-update-pipeline` task in the destroy
pipelines to prevent them from deleting themselves when run. The empty
value, when the env var isn't set, doesn't matter because the pipeline which
interpolate it won't be created.

We have talked about adding a Concourse feature to make the background a
different colour for different environments or adding some text to the
pipeline screen. Those things might still be desirable, but we should leave
this in place for extra safety.